refactor: place more stuff in reducer; cleanup
